scripts/requirements.txt
# Step 5 - Python Dependencies
pandas>=2.1.0
python-dotenv>=1.0.0
supabase>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
